Separate headers into datePickers and dateAxes
